Renamed dir professional-resource to professional.
