+++ webapp/src/app/resource/professional/professional-resource.component.spec.ts
